Fix `SLANG_ENABLE_ASAN` CMake option
Fixes sanitizers silently not being enabled due to broken checks for
compiler support of sanitizer flags.

Fixes non-leak errors reported by ASan when running `slang-test -api
cpu` on Linux.

Fixes #9097.
Fixes #9098.

+++ source/compiler-core/slang-lexer.cpp
@@ -743,7 +743,7 @@ IntegerLiteralValue getIntegerLiteralValue(
         if (digit < 0)
             break;
 
-        value = static_cast<uint64_t>(value) * base + digit;
+        value = value * base + digit;
     }
 
     if (outSuffix)
